fix(test): explicit size_t->double cast in benchmark_simd_all
The last two -Werror=conversion sites in the Strict build:
`(i % 100) * 0.01` implicitly converts size_t -> double. Made the cast
explicit (behavior-preserving). These surfaced only after the OOM fix let
the Strict build compile the full test suite to completion.

+++ tests/benchmark_simd_all.cpp
@@ -65,8 +65,8 @@ class SIMDBenchmark {
           output_simd_(size) {
         // Initialize with varied but safe values
         for (size_t i = 0; i < size; ++i) {
-            input_a_[i] = 1.0 + (i % 100) * 0.01;  // 1.0 to 1.99
-            input_b_[i] = 0.5 + (i % 50) * 0.01;   // 0.5 to 0.99
+            input_a_[i] = 1.0 + static_cast<double>(i % 100) * 0.01;  // 1.0 to 1.99
+            input_b_[i] = 0.5 + static_cast<double>(i % 50) * 0.01;   // 0.5 to 0.99
         }
     }
